Split out util functions to utils package
